web3_py_simple_storage/requirements.txt
black
py-solc-x
python-dotenv